requirements.txt
base58
cryptography
django==3.2
django-environ
PyJWT
requests